debugging. Inlet pumping observerd. Changing ro stag expression to 1.0 not 0.999 removed this. Still some outlet pumping but far less severe and appears to be coming from the corners?

+++ Code/exa_no_rk/types.f90
@@ -23,7 +23,7 @@
           real :: t_tot
           
 !         Frame number stored for time-varying printouts 
-          real :: frame_no
+          integer :: frame_no
 
 !         Second order time derivative factor for Crocco method
           real :: facsec
